1.integration with sse event of reset vault key 2.add notification and remove secrets
